[쇼핑 장바구니] 에디 3,4단계 미션 제출합니다. (#75)
* docs: LiveData + DataBinding = LiveDataBinding???

* docs: ViewModel(아키텍처 패턴) vs ViewModel(AAC)

* chore: gradle room 추가

* feat: 장바구니 db 구현

* feat: 최근 본 상품 db 구현

* feat: 주문 목록 수량 체크

* chore: 로그 제거

* chore: shape으로 원 구현

* feat: 주문 목록 장바구니 배지 추가

* feat: 이전에 장바구니에 있던 상품으로 주문 목록 초기화

* refactor: 장바구니 상품 개수 업데이트 기능

* feat: refactor: 상품 목록으로 왔을 때 장바구니 상품 개수 동기화

* refactor: 장바구니 개수가 0일 때는 배지 표시하지 않기

* refactor: ShoppingCartDatabase createAt 필드 추가

* refactor: product_count visibility 위치 수정

* feat: 상품 상세 페이지에서 상품 개수 산수 기능

* fix: 상품 상세 페이지 장바구니 담기 기능 오류 해결

* feat: 상품 목록에서 최근에 본 상품 보기 기능 구현

* feat: 상품 상세 페이지에서 최근에 본 상품 보기 기능 구현

* fix: 동시성 문제로 인한 상태 오류 해결

* refactor: 상품 목록 페이지 전체 스크롤 가능하게 수정

* refactor: 마지막으로 본 상품을 선택했을 때는 마지막으로 본 상품이 보이지 않게 구현

* chore: gradle okhttp 추가

* feat: MockWebServer 구현

* refacotr: 더미 데이터 이름 변경

* refactor: 최근 및 마지막으로 본 상품에 대한 예외처리

* refactor: MockWebServer 초기화를 백그라운드 스레드에서 처리

* refactor: domain repository 패키지 분리

* refactor: productListPagingSource.load() 성공 결과 함수 파라미터명 수정

* test: 코드 수정으로 인한 테스트 코드 수정

* chore: 주석된 gradle 제거

* fix: 본 상품이 마지막으로 본 상품일 때 발생하는 ui 이슈 해결

* feat: 개수에 따른 총 가격 변화

* test: 테스트 오류 해결

* fix: 상세 페이지에서 개수가 0으로 표시되는 이슈 해결

* refactor: Serializable에서 Parcelable로 개선

* refactor: BaseActivity onCreate 추상화 네임이 수정

* refactor: 최근에 본 상품 데이터를 Product 데이터로 통일

* refactor: 암시적 Intent를 통해 setResult를 하도록 개선

* chore: data model product 네이밍 수정

* refactor: CartProductsAdapter ListAdapter로 개선

* refactor: ProductHistoryListAdapter ListAdapter로 개선

* refactor: ProductListAdapter ListAdapter로 개선

* refactor: local data source 구현

* refactor: remote data source 구현

* feat: 에러 발생시 에러 페이지로 이동

* fix: 필요없는 코드 제거

* fix: 최근 및 마지막 아이템이 안보이는 이슈 해결

* refactor: CartProductsAdapter 아이템 업데이트 로직 간소화

* chore: 로그 삭제

* refactor: application/json 상수화 네이밍 수정

* refactor: UI 상태 유지 개선

* refactor: 인터셉터를 통한 에러 핸들링

* docs: 학습 기록

* chore: 필요없는 코드 제거

+++ app/src/main/java/woowacourse/shopping/presentation/ui/productlist/ProductListActionHandler.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.presentation.ui.productlist
 
 interface ProductListActionHandler {
-    fun navigateToProductDetail(productId: Int)
+    fun navigateToProductDetail(productId: Long)
 
     fun loadMoreProducts()
 }
